Fix the update test.
Confusingly, our mocks here are not acting like real-world and it's making the test hard to write. It starts with a fetchDraftSubmissionMock where we set the ID to be 15. That's because we are going to request id 15 b/c that's what's in the route. However, fetchDraftSubmissionMock isn't smart enough to set the ID of the returned draft to be 14, so we get the default of test-abc-123 instead. The code uses the ID that's in the returned draft to set what is used in the mutation operation, so it ignores the 15 in the mock URL and makes a request using the default ID again.

There are certainly fixes to be made here. Probably by starting with making the ID passed in the mocks what's actually returned in the response. A fix that would hide this would be to use test-abc-123 in the url slug. That would make this test look right but would probably end up tripping us up down the road. I'm leaving it as is for now.