Pin ua-parser-js to 0.7.24 to avoid a poisioned package
Per https://github.com/faisalman/ua-parser-js/issues/536
a set of versions of ua-parser-js were released that had
malware in them. This resolves our version to the version
we had locked at the time of the incident. Eventually
we will want to remove this once the dust has settled
but this prevents us from accidentally installing the
bad package for now.