Only remove the CF stages that exist
Sometimes a CF resource might not exist, so not all of the
expected stages will be there. For example, if someone removes
a stack manually. Or a dangling CF stack from a past failed remove.

This should let us only target the remaining stacks so we don't
get errors attempting to remove resources that don't exist anymore.