Destroy script in TypeScript (#762)
* First pass at using CF SDK for destroy

* refactor func out

* More bucket cleaning

* Refactor a bit

* Remove all versioned objects in s3 buckets

* Comments

* Remove random autoimport

* Setup destroy workflow

* Refactor: getbucketsinstack

* Refactor turnOffVersioningOnBucket

* refactor: get versioned files in bucket

* Refactor for code climate judging me

* Refactor returns

* Only remove the CF stages that exist

Sometimes a CF resource might not exist, so not all of the
expected stages will be there. For example, if someone removes
a stack manually. Or a dangling CF stack from a past failed remove.

This should let us only target the remaining stacks so we don't
get errors attempting to remove resources that don't exist anymore.

* Add an error check to libobjectversions

* Fix up error messaging and exit status

* Less noise in output

* Start pushing some of the errors up

* Remove the bash

* Some unneeded code cleaning

* Make sure we split on large bucket clears

* Protect some stages, just in case

* Noise